twitter-scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.3
python-dotenv==1.0.0
supabase>=2.24.0
textblob==0.17.1
lxml==5.1.0
langdetect==1.0.9
fasttext==0.9.2
langid==1.1.6
httpx>=0.27.2
websockets>=15.0

